fix clusterrolebinding overlap
ClusterRoleBinding has no namespace, but binds to a service account
in a given namespace.

Each instance of tezos-on-gke in a cluster overwrites the same
ClusterRoleBinding, resulting in the last deployment "erasing" the
previous one.

When the cluster role binding gets erased, prometheus in the affected
namespace can no longer scrape its targets, resulting in alerting being
broken.

We fix by appending the namespace name to the cluster role binding.